wnioskiurlopowe-django/requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.9
django-crispy-forms==1.13.0
django-environ==0.8.1
django-heroku==0.3.1
django-model-utils==4.2.0
gunicorn==20.1.0
Pillow==8.4.0
psycopg2==2.9.1
python-decouple==3.5
pytz==2021.3
reportlab==3.6.2
sqlparse==0.4.2
Unipath==1.1
whitenoise==5.3.0